\name{AnnotationAM}
\docType{data}
\alias{AnnotationAM}
\title{Annotation of proteins by cell fraction}
\description{
  Annotation of proteins collected from Metadata.
}
\usage{
AnnotationAM}
\format{
  data.frame

  [,1] protein ID (uniprot ID)

  [,2] protein ID (genename)

  [,3] annotations - Christoforou

  [,4] annotations - Uniprot

  [,5] annotations - G0

  [,6] annotations - HPA supportive

  [,7] annotations - HPA uncertain

  [,8] annotations - Christoforu and Uniprot/G0 overlap

  [,9] annotations - Uniprot G0 SUM

  [,10] annotations - Uniprot G0 overlap
}
