\name{AnnotationAM}
\docType{data}
\alias{AnnotationAM}
\title{Annotation of proteins by cell fraction}
\description{
  Annotation of proteins collected from Metadata.
}
\usage{
AnnotationAM}
\format{
  data.frame

  [,1] protein ID (uniprot ID)

  [,2] protein ID (genename)

  [,3] annotations - Christoforou Uniprot G.O overlap

  [,4] annotations - Christoforou Uniprot G.O sum

  [,5] annotations - HPA Single supportive

  [,6] annotations - HPA Single uncertain

  [,7] annotations - HPA main supportive

  [,8] annotations - HPA main uncertain

  [,9] annotations - Christoforou Uniprot G.O overlap(CMN)
}
