\name{AnnotationAM}
\docType{data}
\alias{AnnotationAM}
\title{Annotation of proteins by cell fraction}
\description{
  Annotation of proteins collected from Metadata.
}
\usage{
AnnotationAM}
\format{
  data.frame

  [,1] protein ID (uniprot ID)

  [,2] protein ID (genename)

  [,3] annotations - Christoforou

  [,4] annotations - Uniprot

  [,5] annotations - G 0

  [,6] annotations - HPA supportive

  [,7] annotations - HPA uncertain
}
